+++ modulo6/classes-encapsulamento/template-classes-encapsulamento/src/index.ts
@@ -1,6 +1,4 @@
-import express from 'express'
-import cors from 'cors'
-import dotenv from 'dotenv'
+import app from "./app"
 import { ping } from './endpoints/ping'
 import { createUser } from './endpoints/createUser'
 import { getUsers } from './endpoints/getUsers'
@@ -9,15 +7,6 @@ import { getProducts } from './endpoints/getProducts'
 import { createPurchase } from './endpoints/createPurchase'
 import { getUserPurchases } from './endpoints/getUserPurchases'
 
-dotenv.config()
-const app = express()
-
-app.use(express.json())
-app.use(cors())
-
-app.listen(process.env.PORT || 3003, () => {
-  console.log(`Servidor rodando na porta ${process.env.PORT || 3003}`)
-})
 
 app.get("/ping", ping)
